Performance boost for Problem 78 Coin partitions, no-longer slow.

+++ net.projecteuler.solutions/src/main/java/problem/no71to80/Problem78.java
@@ -1,9 +1,7 @@
 package problem.no71to80;
 
-import java.math.BigInteger;
-
 import static problem.Solution.problem;
-import static utils.sequence.arithmetic.A046641.a;
+import static utils.sequence.arithmetic.A046641Fast.a;
 
 public class Problem78 {
     public static void main(String[] args) {
@@ -14,7 +12,7 @@ public class Problem78 {
     private static final int ONE_MILLION = 1_000_000;
 
     static int leastValueOfCoinsWhereCombinationsDivisibleBy1_000_000() {
-        return a(BigInteger.valueOf(ONE_MILLION));
+        return a(ONE_MILLION);
     }
 
     private Problem78() {}
